Moved PlatformMutex/PlatformMonitor
