+++ src/hotspot/share/memory/metaspace/testHelpers.cpp
@@ -29,6 +29,7 @@
 #include "memory/metaspace/metaspaceContext.hpp"
 #include "memory/metaspace/testHelpers.hpp"
 #include "runtime/mutexLocker.hpp"
+#include "runtime/os.hpp"
 #include "utilities/debug.hpp"
 #include "utilities/globalDefinitions.hpp"
 #include "utilities/ostream.hpp"
